Add files via upload
A sketch of how the model looks like.
On the left side there is the "pricing" branch: stock information are put into a Stochastic Discount Factor (SD). Weights in the SDF are updated so that the loss function minimizes pricing error.
On the right side, test assets are selected. The adversarial update tries to select assets so that loss is maximized.